pyproject.toml
[tool.poetry]
name = "ninja_assistant"
version = "0.1.0"
description = ""
authors = ["DJWOMS <socanime@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-ninja = "^0.13.2"
uvicorn = "^0.14.0"
Pillow = "^8.3.1"
psycopg2 = "^2.9.1"
PyJWT = "^2.1.0"
google-auth = "^1.33.0"
requests = "^2.26.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
